@@ -1,18 +1,23 @@
-# coding: utf-8
-
 class ApplicationController < ActionController::Base
-  protect_from_forgery
-  
-  layout 'testeroni'
-  
+
+  before_action :configure_permitted_parameters, if: :devise_controller?
+
   USER, PASSWORD = 'testeroni', 'besteroni'
-  
-  before_filter :setup_user
-  before_filter :set_return_to, :except => ['sign_in']
+
+  # before_filter :setup_user
+  before_action :set_return_to, :except => ['sign_in']
   # before_filter :promo
   # before_filter :auth_in_prod
   # before_filter :log_session
-  
+
+  protected
+
+  def configure_permitted_parameters
+    # added_attrs = [:username, :email, :password, :password_confirmation, :remember_me]
+    devise_parameter_sanitizer.permit :sign_up, keys: [:name]
+    devise_parameter_sanitizer.permit :account_update, keys: [:name, :surname, :email, :avatar]
+  end
+
   def auth_in_prod
     if Rails.env == 'production'
       authenticate_or_request_with_http_basic do |user, password|
@@ -20,23 +25,23 @@ class ApplicationController < ActionController::Base
       end
     end
   end
-  
+
   # application_controller before_filter
   def setup_user
     #for hidden signup form
     @user = User.new(:email_list => true) unless user_signed_in?
   end
-  
+
   # application_controller before_filter
   def set_return_to
-    unless request.referrer.include?('auth') || request.path.include?('questions') || request.path.include?('users')
+    unless request.referrer.present? && (request.referrer.include?('auth') || request.path.include?('questions') || request.path.include?('users'))
       logger.debug "set_return_to request.path: #{request.path}"
       session[:return_to] = request.path
     else
       logger.debug "DID NOT set_return_to request.path: #{request.path}"
     end
   end
-  
+
   # adapted from: http://groups.google.com/group/plataformatec-devise/tree/browse_frm/month/2010-06?_done=/group/plataformatec-devise/browse_frm/month/2010-06%3F&
   def stored_location_for(resource)
     if current_user
@@ -58,7 +63,7 @@ class ApplicationController < ActionController::Base
     end
     super( resource )
   end
-  
+
   # rails generate model Activity actor_id:integer, verb:integer, object_type:string, object_id:integer
   def record_activity(user, verb, object)
     if object.class == Tst
@@ -68,7 +73,7 @@ class ApplicationController < ActionController::Base
     end
     Activity.create(:user_id => user.id, :verb => verb, :object_type => otype, :object_id => object.id)
   end
-  
+
   def promo
     if promo = true
       msg = "<span style='color:red;'>ALPHA ALERT!</span> This is a test server, don't rely on data sticking around (but <a href='mailto:john@wordie.org'>email me</a> if it needs to)"
@@ -80,15 +85,15 @@ class ApplicationController < ActionController::Base
       clear_promo
     end
   end
-  
+
   def log_session
     logger.debug '--'
     logger.debug "session: #{session.inspect}"
     logger.debug "user_session: #{user_session.inspect}"
     logger.debug '--'
   end
-  
-  
+
+
   # manages the session info for promo. don't call this directly. (see get_promo)
   def set_promo(msg, display_after)
     session[:promo] ||= {}
@@ -101,7 +106,7 @@ class ApplicationController < ActionController::Base
   def clear_promo
     session[:promo] = @promo = nil
   end
-  
+
   # IF a user has taken tests but wasn't logged in, they'll be in the session. Check for those and persist them if they're there.
   def save_take_and_set_flash(user)
     if session[:take]
@@ -111,16 +116,16 @@ class ApplicationController < ActionController::Base
       flash[:notice] = "Signed in successfully."
     end
   end
-  
+
   def url_escape(val)
     CGI.escape(val).gsub('+', '%20')
   end
   helper_method :url_escape
-  
+
   protected
   def after_sign_up_path_for(resource)
     logger.debug "GOT TO after_sign_up_path_for in application_controller"
     'http://www.google.com'
   end
-  
+
 end
