@@ -1,8 +1,8 @@
-class Follow < ActiveRecord::Base
-  
+class Follow < ApplicationRecord
+
   # objects
   USER = 0
   TST = 1
   COMMENT = 2
-  
+
 end
