@@ -1,6 +1,4 @@
-# coding: utf-8
-
-class Vote < ActiveRecord::Base
+class Vote < ApplicationRecord
 
   scope :for_voter, lambda { |*args| where(["voter_id = ? AND voter_type = ?", args.first.id, args.first.class.name]) }
   scope :for_voteable, lambda { |*args| where(["voteable_id = ? AND voteable_type = ?", args.first.id, args.first.class.name]) }
