@@ -1,18 +1,18 @@
-class CreateSlugs < ActiveRecord::Migration
+class CreateSlugs < ActiveRecord::Migration[6.0]
   def self.up
-    create_table :slugs do |t|
-      t.string :name
-      t.integer :sluggable_id
-      t.integer :sequence, :null => false, :default => 1
-      t.string :sluggable_type, :limit => 40
-      t.string :scope
-      t.datetime :created_at
-    end
-    add_index :slugs, :sluggable_id
-    add_index :slugs, [:name, :sluggable_type, :sequence, :scope], :name => "index_slugs_on_n_s_s_and_s", :unique => true
+    # create_table :slugs do |t|
+    #   t.string :name
+    #   t.integer :sluggable_id
+    #   t.integer :sequence, :null => false, :default => 1
+    #   t.string :sluggable_type, :limit => 40
+    #   t.string :scope
+    #   t.datetime :created_at
+    # end
+    # add_index :slugs, :sluggable_id
+    # add_index :slugs, [:name, :sluggable_type, :sequence, :scope], :name => "index_slugs_on_n_s_s_and_s", :unique => true
   end
 
   def self.down
-    drop_table :slugs
+    # drop_table :slugs
   end
 end
