@@ -1,31 +1,30 @@
+# frozen_string_literal: true
+
 class Users::OmniauthCallbacksController < Devise::OmniauthCallbacksController
-  
-  def twitter
-    @user = User.find_for_twitter_oauth(env["omniauth.auth"], current_user, cookies[:email])
+  # You should configure your model like this:
+  # devise :omniauthable, omniauth_providers: [:twitter]
 
-    if @user.persisted?
-      save_take_and_set_flash(@user)
-      
-      flash[:notice] = I18n.t "devise.omniauth_callbacks.success", :kind => "Twitter"
-      sign_in_and_redirect @user, :event => :authentication
-    else
-      session["devise.facebook_data"] = env["omniauth.auth"]
-      redirect_to new_user_registration_url
-    end
-  end
-  
-  def facebook
-    @user = User.find_for_facebook_oauth(env["omniauth.auth"], current_user)
+  # You should also create an action method in this controller like this:
+  # def twitter
+  # end
 
-    if @user.persisted?
-      save_take_and_set_flash(@user)
-      
-      flash[:notice] = I18n.t "devise.omniauth_callbacks.success", :kind => "Facebook"
-      sign_in_and_redirect @user, :event => :authentication
-    else
-      session["devise.facebook_data"] = env["omniauth.auth"]
-      redirect_to new_user_registration_url
-    end
-  end
-  
+  # More info at:
+  # https://github.com/plataformatec/devise#omniauth
+
+  # GET|POST /resource/auth/twitter
+  # def passthru
+  #   super
+  # end
+
+  # GET|POST /users/auth/twitter/callback
+  # def failure
+  #   super
+  # end
+
+  # protected
+
+  # The path used when OmniAuth fails
+  # def after_omniauth_failure_path_for(scope)
+  #   super(scope)
+  # end
 end
