@@ -1,5 +1,5 @@
 # coding: utf-8
 
 module TagsHelper
-  include ActsAsTaggableOn::TagsHelper
+  # include ActsAsTaggableOn::TagsHelper
 end
